python_requirements.txt
google-generativeai>=0.3.0
Pillow>=9.0.0